# services/intersection-api/docs/swagger-docs/openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [{ "url": "http://localhost:8089", "description": "Generated server url" }],
  "paths": {
    "/decoder/upload": {
      "post": {
        "tags": ["decoder-controller"],
        "operationId": "decode_request",
        "parameters": [
          { "name": "test", "in": "query", "required": false, "schema": { "type": "boolean", "default": false } }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EncodedMessage" } } },
          "required": true
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "string" } } } }
        }
      }
    },
...
    "/map/json": {
      "get": {
        "tags": ["map-controller"],
        "operationId": "findMaps",
        "parameters": [
          {
            "name": "intersection_id",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "start_time_utc_millis",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "end_time_utc_millis",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int64" }
          },
          { "name": "latest", "in": "query", "required": false, "schema": { "type": "boolean", "default": false } },
          { "name": "test", "in": "query", "required": false, "schema": { "type": "boolean", "default": false } },
          { "name": "compact", "in": "query", "required": false, "schema": { "type": "boolean", "default": false } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProcessedMapLineString" } }
              }
            }
          }
        }
      }
    },
    "/map/count": {
      "get": {
        "tags": ["map-controller"],
        "operationId": "countMaps",
        "parameters": [
          {
            "name": "intersection_id",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "start_time_utc_millis",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "end_time_utc_millis",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int64" }
          },
          { "name": "test", "in": "query", "required": false, "schema": { "type": "boolean", "default": false } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "type": "integer", "format": "int64" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EncodedMessage": {
        "type": "object",
        "properties": {
          "asn1Message": { "type": "string" },
          "type": { "type": "string", "enum": ["UNKNOWN", "BSM", "MAP", "SPAT", "SRM", "SSM", "TIM"] }
        }
      },
...
      "ProcessedValidationMessage": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "jsonPath": { "type": "string" },
          "schemaPath": { "type": "string" },
          "exception": { "type": "string" }
        }
      },
...
      "ConnectingLanesFeatureCollectionLineString": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "features": { "type": "array", "items": { "$ref": "#/components/schemas/ConnectingLanesFeatureLineString" } }
        }
      },
      "ConnectingLanesFeatureLineString": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "id": { "type": "string" },
          "geometry": { "$ref": "#/components/schemas/LineString" },
          "properties": { "$ref": "#/components/schemas/ConnectingLanesProperties" }
        }
      },
      "ConnectingLanesProperties": {
        "type": "object",
        "properties": {
          "signalGroupId": { "type": "integer", "format": "int32" },
          "ingressLaneId": { "type": "integer", "format": "int32" },
          "egressLaneId": { "type": "integer", "format": "int32" }
        }
      },
      "J2735ConnectingLane": {
        "type": "object",
        "properties": {
          "lane": { "type": "integer", "format": "int32" },
          "maneuver": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          }
        }
      },
      "J2735Connection": {
        "type": "object",
        "properties": {
          "connectingLane": { "$ref": "#/components/schemas/J2735ConnectingLane" },
          "remoteIntersection": { "$ref": "#/components/schemas/J2735IntersectionReferenceID" },
          "signalGroup": { "type": "integer", "format": "int32" },
          "userClass": { "type": "integer", "format": "int32" },
          "connectionID": { "type": "integer", "format": "int32" }
        }
      },
      "J2735IntersectionReferenceID": {
        "type": "object",
        "properties": {
          "region": { "type": "integer", "format": "int32" },
          "id": { "type": "integer", "format": "int32" }
        }
      },
      "J2735LaneTypeAttributes": {
        "type": "object",
        "properties": {
          "vehicle": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "crosswalk": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "bikeLane": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "sidewalk": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "median": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "striping": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "trackedVehicle": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "parking": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          }
        }
      },
      "J2735RegulatorySpeedLimit": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "unknown",
              "maxSpeedInSchoolZone",
              "maxSpeedInSchoolZoneWhenChildrenArePresent",
              "maxSpeedInConstructionZone",
              "vehicleMinSpeed",
              "vehicleMaxSpeed",
              "vehicleNightMaxSpeed",
              "truckMinSpeed",
              "truckMaxSpeed",
              "truckNightMaxSpeed",
              "vehiclesWithTrailersMinSpeed",
              "vehiclesWithTrailersMaxSpeed",
              "vehiclesWithTrailersNightMaxSpeed"
            ]
          },
          "speed": { "type": "number" }
        }
      },
      "LineString": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "coordinates": {
            "type": "array",
            "items": { "type": "array", "items": { "type": "number", "format": "double" } }
          },
          "bbox": { "type": "array", "items": { "type": "number", "format": "double" } }
        }
      },
      "MapFeatureCollectionLineString": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "features": { "type": "array", "items": { "$ref": "#/components/schemas/MapFeatureLineString" } }
        }
      },
      "MapFeatureLineString": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "id": { "type": "integer", "format": "int32" },
          "geometry": { "$ref": "#/components/schemas/LineString" },
          "properties": { "$ref": "#/components/schemas/MapProperties" }
        }
      },
      "MapNode": {
        "type": "object",
        "properties": {
          "delta": { "type": "array", "items": { "type": "integer", "format": "int32" } },
          "stopLine": { "type": "boolean" },
          "dwidth": { "type": "integer", "format": "int32" },
          "delevation": { "type": "integer", "format": "int32" }
        }
      },
      "MapProperties": {
        "type": "object",
        "properties": {
          "nodes": { "type": "array", "items": { "$ref": "#/components/schemas/MapNode" } },
          "laneId": { "type": "integer", "format": "int32" },
          "laneName": { "type": "string" },
          "laneType": { "$ref": "#/components/schemas/J2735LaneTypeAttributes" },
          "sharedWith": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "egressApproach": { "type": "integer", "format": "int32" },
          "ingressApproach": { "type": "integer", "format": "int32" },
          "ingressPath": { "type": "boolean" },
          "egressPath": { "type": "boolean" },
          "maneuvers": {
            "type": "object",
            "properties": { "empty": { "type": "boolean" } },
            "additionalProperties": { "type": "boolean" }
          },
          "connectsTo": { "type": "array", "items": { "$ref": "#/components/schemas/J2735Connection" } }
        }
      },
      "MapSharedProperties": {
        "type": "object",
        "properties": {
          "schemaVersion": { "type": "integer", "format": "int32" },
          "messageType": { "type": "string" },
          "odeReceivedAt": { "type": "string", "format": "date-time" },
          "originIp": { "type": "string" },
          "intersectionName": { "type": "string" },
          "region": { "type": "integer", "format": "int32" },
          "intersectionId": { "type": "integer", "format": "int32" },
          "msgIssueRevision": { "type": "integer", "format": "int32" },
          "revision": { "type": "integer", "format": "int32" },
          "refPoint": { "$ref": "#/components/schemas/OdePosition3D" },
          "cti4501Conformant": { "type": "boolean" },
          "validationMessages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProcessedValidationMessage" }
          },
          "laneWidth": { "type": "integer", "format": "int32" },
          "speedLimits": { "type": "array", "items": { "$ref": "#/components/schemas/J2735RegulatorySpeedLimit" } },
          "mapSource": { "type": "string", "enum": ["RSU", "V2X", "MMITSS", "unknown"] },
          "timeStamp": { "type": "string", "format": "date-time" }
        }
      },
      "OdePosition3D": {
        "type": "object",
        "properties": {
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "elevation": { "type": "number" }
        }
      },
      "ProcessedMapLineString": {
        "type": "object",
        "properties": {
          "mapFeatureCollection": { "$ref": "#/components/schemas/MapFeatureCollectionLineString" },
          "connectingLanesFeatureCollection": {
            "$ref": "#/components/schemas/ConnectingLanesFeatureCollectionLineString"
          },
          "properties": { "$ref": "#/components/schemas/MapSharedProperties" }
        }
      },
...
    }
  }
}